Modify context handler into session handler
